backend/requirements.txt
Django==2.1.7
django-polymorphic==2.0.3
pytz==2018.9
